+++ src/main/java/org/betonquest/betonquest/quest/event/weather/Weather.java
@@ -39,10 +39,14 @@ public enum Weather {
     /**
      * Apply the weather to the given world.
      *
-     * @param world world to change the weather in
+     * @param world    world to change the weather in
+     * @param duration how long the weather will remain - values &lt;= 0 won't set a duration
      */
-    public void applyTo(final World world) {
+    public void applyTo(final World world, final int duration) {
         world.setStorm(storm);
         world.setThundering(thunder);
+        if (duration > 0) {
+            world.setWeatherDuration(duration);
+        }
     }
 }